Reorder precision specifiers.

+++ src/x_ite/Components/X_ITE/GaussianSplats.js
@@ -20,9 +20,9 @@ import Matrix4              from "../../../standard/Math/Numbers/Matrix4.js";
 const vs = () => /* glsl */ `#version 300 es
 
 precision highp int;
-precision mediump float;
-precision mediump sampler2D;
-precision mediump sampler2DArray;
+precision highp float;
+precision highp sampler2D;
+precision highp sampler2DArray;
 
 uniform ivec4 x3d_Viewport;
 uniform mat4  x3d_ProjectionMatrix;
@@ -331,8 +331,8 @@ main ()
 const fs = () => /* glsl */ `#version 300 es
 
 precision highp int;
-precision mediump float;
-precision mediump sampler2D;
+precision highp float;
+precision highp sampler2D;
 
 in vec4 color;
 in vec2 texCoord;
